exams/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.5
django-filter==23.2
django-pgtrigger==4.7.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.4
Faker==12.0.1
flake8==6.1.0
flake8-isort==6.0.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mixer==7.2.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
